Dedicated file for helpers & test for buildFileTargetPath
