--- packages/core/cli/src/scripts/createProject/buildFileTargetPath.ts
@@ -1,1 +1,0 @@
-export default (file: string, targetPath: string, chopPhrase: string): string => targetPath + (file.replace(chopPhrase, ''));
