--- packages/core/cli/src/scripts/createProject/getThemePath.ts
@@ -1,1 +1,0 @@
-export default (themeName: string): string => `../../../node_modules/@vue-storefront/${themeName}`;
