+++ packages/core/cli/src/utils/helpers.ts
@@ -1,0 +1,2 @@
+export const buildFileTargetPath = (file: string, targetPath: string, chopPhrase: string): string => targetPath + (file.replace(chopPhrase, ''));
+export const getThemePath = (themeName: string): string => `../../../node_modules/@vue-storefront/${themeName}`;
